Moved view related components to their own files for edit mode
