+++ client/src/components/view/LoadPropertyOptions.tsx
@@ -1,0 +1,37 @@
+import { useEffect } from 'react';
+import { useAppDispatch, useAppSelector } from '../../app/hooks';
+import { useGetCategoriesQuery } from '../../features/apiSlice';
+import { setPropertyOptions } from './createViewSlice';
+
+// TODO: Api endpoint to get all property options
+
+export const LoadPropertyOptions = () => {
+  const module = useAppSelector((state) => state.createView.module);
+  const { data: inventory = [] } = useGetCategoriesQuery('inventory');
+  const { data: crm = [] } = useGetCategoriesQuery('crm');
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    const getPropertyOptions = (): string[] => {
+      const propertyOptions: string[] = [];
+
+      for (const category of inventory) {
+        const properties = Object.keys(category.itemShape);
+        propertyOptions.push(...properties);
+      }
+
+      for (const category of crm) {
+        const properties = Object.keys(category.itemShape);
+        propertyOptions.push(...properties);
+      }
+
+      const uniques = Array.from(new Set(propertyOptions));
+      return uniques.sort((a, b) => a.localeCompare(b));
+    };
+
+    const properties = getPropertyOptions();
+    dispatch(setPropertyOptions({ properties }));
+  }, [inventory, crm, module, dispatch]);
+
+  return null;
+};
